+++ src/db/mod.rs
@@ -1,0 +1,2 @@
+pub mod user_db;
+pub mod marks_db;
