+++ src/handlers/mod.rs
@@ -1,0 +1,11 @@
+pub mod email;
+pub mod user;
+pub mod auth;
+pub mod password;
+pub mod marks;
+
+pub use email::send_confirmation_email;
+pub use user::create_user;
+pub use auth::login_user;
+pub use password::forgot_password;
+pub use marks::add_marks;
