+++ src/main.rs
@@ -17,6 +17,7 @@ async fn main() -> std::io::Result<()> {
             .route("/login", web::post().to(handlers::login_user))     // Login user
             .route("/forgot-password", web::post().to(handlers::forgot_password)) // Forgot password
             .route("/add-marks", web::post().to(handlers::add_marks)) // Add marks endpoint
+            // .route("/edit-mark", web::post().to(handlers::edit_marks::edit_mark)) // Edit marks route
     })
     .bind("127.0.0.1:8081")?
     .run()
