--- src/models.rs
@@ -1,38 +1,0 @@
-use mongodb::bson::oid::ObjectId;
-use serde::{Deserialize, Serialize};
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct User {
-    #[serde(rename = "_id", skip_serializing_if = "Option::is_none")]
-    pub id: Option<ObjectId>,
-    pub username: String,
-    pub email: String,
-    pub password: String,
-    pub user_type: String, // User type: student, teacher, admin
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct UserLogin {
-    #[serde(rename = "_id", skip_serializing_if = "Option::is_none")]
-    pub id: Option<ObjectId>,
-    pub username: String,
-    pub password: String,
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct PasswordChangeRequest {
-    pub username: String,
-    pub old_password: String,
-    pub new_password: String,
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct Marks {
-    #[serde(rename = "_id", skip_serializing_if = "Option::is_none")]
-    pub id: Option<ObjectId>,
-    pub student_id: String,
-    pub subject1: i32,
-    pub subject2: i32,
-    pub teacher_name: String, // Teacher ka naam
-    pub user_type: String, // User type to check if the user is a teacher
-}
