+++ src/models/mod.rs
@@ -1,0 +1,10 @@
+pub mod user;
+pub mod login;
+pub mod password_change;
+pub mod marks;
+pub mod edit_mark;
+
+pub use user::User;
+pub use login::UserLogin;
+pub use password_change::PasswordChangeRequest;
+pub use marks::Marks;
