+++ src/models/password_change.rs
@@ -1,0 +1,8 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Serialize, Deserialize, Clone)]
+pub struct PasswordChangeRequest {
+    pub username: String,
+    pub old_password: String,
+    pub new_password: String,
+}
